+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [CalVer](https://calver.org/).
 
 ## [Unreleased]
 
+### Changed
+
+- Data: keio-bus の GTFS resource を 20260724 版へ更新。
+- Data: nishi-tokyo-bus の GTFS resource を 20260718 版へ更新。
+- Data: kyoto-bus の GTFS resource を 20260724 版へ更新。
+
 ## [2026.07.17]
 
 ### Changed